requirements.txt
fastapi
httpx
motor
python-dotenv